# api/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Source Reconciliation API
  version: "0.1.0"
  description: >
    Contract for the Razorpay multi-source reconciliation demo (spec §12).
    Paths and schemas only — no implementation. Every response schema mirrors
    the frozen pydantic contracts in core/models.py; do not change field
    names, types, or nullability here without re-syncing that file.
paths:
...
  /api/auth/login:
    post:
      operationId: login
      summary: Exchange a seeded credential for a session cookie
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: >
            Authenticated. The session travels in a `Set-Cookie:
            recon_session=...; HttpOnly; SameSite=Lax` header; the body names
            the org so a client need not decode the cookie to know which tenant
            it is looking at.
          headers:
            Set-Cookie:
              description: The signed session cookie. HttpOnly.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: >
            The credential was not accepted. **One response for both failure
            modes**: an unknown address and a wrong password are identical in
            status, body and latency, because answering them differently is an
            account-enumeration oracle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
        "503":
          $ref: "#/components/responses/AuthDisabled"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: recon_session
      description: >
        The signed session cookie issued by POST /api/auth/login. Signed with
        HMAC-SHA256 over RECON_SECRET_KEY -- not encrypted, and not a JWT: the
        payload is three server-issued claims (subject, org, expiry) and no
        algorithm is negotiated with the client.

        Deliberately NOT declared as a global `security:` requirement. Whether
        it is required is a deployment decision (RECON_AUTH), and a contract
        that asserted "always required" would be false for the configuration
        this system is normally run in. Each operation instead declares the 401
        it can answer.
...
  responses:
...
    AuthDisabled:
      description: >
        Authentication is not enabled on this deployment, so there is no
        session to issue. 503 rather than 404 on purpose: the operation exists
        and this contract declares it; what is absent is configuration, and the
        detail names the variable to set.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ServiceUnavailableError"
...
  schemas:
...
    LoginRequest:
      type: object
      description: >
        The whole of this product's identity input surface. There is no signup,
        no password reset and no invite: accounts are seeded from the
        RECON_USERS environment variable, which makes adding one a deployment
        action. That is a scope decision, a recorded scope decision rather than
        half-built here.
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
        password:
          type: string
          description: Never logged, never echoed, never persisted.
    LoginResponse:
      type: object
      description: >
        Who the session says you are. The org id is returned so a client can
        label its own screen without decoding the cookie -- it is not an input
        anywhere, and no request body on this API accepts one.
      required: [email, org_id]
      additionalProperties: false
      properties:
        email:
          type: string
        org_id:
          type: string
    UnauthorizedError:
      type: object
      description: >
        FastAPI's HTTPException shape, for a 401. Same shape as NotFoundError
        and named separately so the two carry their own descriptions.
      required: [detail]
      properties:
        detail:
          type: string
          description: >
            Names the missing thing -- a session -- and the route that issues
            one. Never names a user, an org, or the value of any secret.
    ServiceUnavailableError:
      type: object
      description: FastAPI's HTTPException shape, for a 503.
      required: [detail]
      properties:
        detail:
          type: string
          description: Names the environment variable that is not set.
...